IOS-4104 Move space edit within settings screen
