+++ Anytype/Resources/Strings/en.lproj/Localizable.strings
@@ -220,6 +220,7 @@
 // Settings
 "Wallpaper" = "Wallpaper";
 "Log out" = "Log out";
+"Misc" = "Misc";
 
 // Action
 "Restore" = "Restore";
