+++ Anytype/Sources/PresentationLayer/Common/SwiftUI/Buttons/RoundedButton.swift
@@ -15,12 +15,14 @@ enum RoundedButtonDecoration {
 struct RoundedButton: View {
     
     let text: String
+    let textColor: Color
     let icon: ImageAsset?
     let decoration: RoundedButtonDecoration?
     let action: () -> Void
     
-    init(text: String, icon: ImageAsset? = nil, decoration: RoundedButtonDecoration? = nil, action: @escaping () -> Void) {
+    init(_ text: String, textColor: Color = .Text.primary, icon: ImageAsset? = nil, decoration: RoundedButtonDecoration? = nil, action: @escaping () -> Void) {
         self.text = text
+        self.textColor = textColor
         self.icon = icon
         self.decoration = decoration
         self.action = action
@@ -36,7 +38,7 @@ struct RoundedButton: View {
                         .frame(width: 24, height: 24)
                     Spacer.fixedWidth(8)
                 }
-                AnytypeText(text, style: .previewTitle1Regular)
+                AnytypeText(text, style: .previewTitle1Regular).foregroundColor(textColor)
                 Spacer()
                 decorationView
                 Spacer.fixedWidth(8)
@@ -67,5 +69,5 @@ struct RoundedButton: View {
 }
 
 #Preview {
-    RoundedButton(text: "ClickMe", icon: .X24.member, decoration: .badge(445)) { }
+    RoundedButton("ClickMe", icon: .X24.member, decoration: .badge(445)) { }
 }
