+++ Anytype/Sources/PresentationLayer/Flows/SpaceHub/SpaceHubCoordinatorView.swift
@@ -93,8 +93,6 @@ struct SpaceHubCoordinatorView: View {
                             switch data {
                             case .mainScreen(let info):
                                 NewSpaceSettingsCoordinatorView(workspaceInfo: info)
-                            case .spaceDetails(let info):
-                                SpaceDetailsView(info: info)
                             }
                         }
                      }
