+++ Anytype/Sources/PresentationLayer/SpaceSettings/SpaceSettings/NewSpaceSettingsModuleOutput.swift
@@ -2,7 +2,6 @@ import Foundation
 
 @MainActor
 protocol NewSpaceSettingsModuleOutput: AnyObject {
-    func onSpaceDetailsSelected()
     func onWallpaperSelected()
     func onDefaultObjectTypeSelected()
