+++ Anytype/Sources/PresentationLayer/TextEditor/Assembly/SettingsScreenData.swift
@@ -3,7 +3,6 @@ import Services
 enum SettingsScreenData: Hashable, Codable, Equatable, Identifiable {
     // Anytype widget screens
     case mainScreen(info: AccountInfo)
-    case spaceDetails(info: AccountInfo)
     
     var id: Int {
         hashValue
@@ -12,8 +11,6 @@ enum SettingsScreenData: Hashable, Codable, Equatable, Identifiable {
     var spaceId: String {
         switch self {
         case .mainScreen(let info):
-            return info.accountSpaceId
-        case .spaceDetails(let info):
             return info.accountSpaceId
         }
     }
